+++ NEWS.md
@@ -1,7 +1,8 @@
 # anthroplus (development version)
 
 * Increased the upper age limit to 229 months exclusive
-  from 228 months inclusive.
+  from 228 months inclusive. Also the upper age limit of 121 months
+  for WFA is now exclusive.
 
 # anthroplus 1.0.0