+++ NEWS.md
@@ -3,6 +3,7 @@
 * Increased the upper age limit to 229 months exclusive
   from 228 months inclusive. Also the upper age limit of 121 months
   for WFA is now exclusive.
+* Changed the maintainer of the package.
 
 # anthroplus 1.0.0